Migrate depth min/max services to float for fractional Stereo ace depth
The Stereo ace BslDepthMinDepth/MaxDepth nodes are IFloatEx (meters), while
blaze and Stereo mini DepthMin/Max are IIntegerEx (mm). The previous int-only
depth chain truncated the ace to whole meters. Migrate the shared depth chain
to double/float32 so the ace accepts fractional meters; blaze/mini round to
their integer mm nodes via std::llround.

Breaking change: set_depth_min and set_depth_max now use SetFloatValue instead
of SetIntegerValue; CurrentParams.depth_min/depth_max are now float32.

Validated on hardware: ace accepts fractional meters (range [0.1, 100] m);
blaze rounds 10.5 -> 11 mm and reports success.

+++ pylon_ros2_camera_component/include/internal/impl/pylon_ros2_camera_3d.hpp
@@ -142,8 +142,8 @@ public:
 
     // Returns the working depth range (mm) by reading the DepthMin/DepthMax nodes
     // from the device node map; returns -1 when the nodes are not available.
-    virtual int getDepthMin() override;
-    virtual int getDepthMax() override;
+    virtual double getDepthMin() override;
+    virtual double getDepthMax() override;
 
     // Returns the current gain/gamma by reading the device node map of the active
     // camera; returns -1 when the node is missing or the read fails. The base
@@ -387,7 +387,7 @@ bool PylonROS23DCamera::isCamRemoved()
     }
 }
 
-int PylonROS23DCamera::getDepthMin()
+double PylonROS23DCamera::getDepthMin()
 {
     int min_depth = -1, max_depth = -1;
     try
@@ -396,12 +396,12 @@ int PylonROS23DCamera::getDepthMin()
     }
     catch (const GenICam::GenericException&)
     {
-        return -1;
+        return -1.0;
     }
-    return min_depth;
+    return static_cast<double>(min_depth);
 }
 
-int PylonROS23DCamera::getDepthMax()
+double PylonROS23DCamera::getDepthMax()
 {
     int min_depth = -1, max_depth = -1;
     try
@@ -410,9 +410,9 @@ int PylonROS23DCamera::getDepthMax()
     }
     catch (const GenICam::GenericException&)
     {
-        return -1;
+        return -1.0;
     }
-    return max_depth;
+    return static_cast<double>(max_depth);
 }
 
 float PylonROS23DCamera::currentGain()
